+++ src/index.ts
@@ -8,6 +8,8 @@ export { isString } from "./lib/types/isString";
 export { isNumber } from "./lib/types/isNumber";
 export { isBoolean } from "./lib/types/isBoolean";
 export { isExtendable } from "./lib/types/isExtendable";
+export { isRegExp } from "./lib/types/isRegExp";
+export { isDate } from "./lib/types/isDate";
 
 // matchers
 export { isSoftMatch } from "./lib/matchers/isSoftMatch";