+++ src/lib/types/index.ts
@@ -9,3 +9,5 @@ export { isString } from "./isString";
 export { isNumber } from "./isNumber";
 export { isBoolean } from "./isBoolean";
 export { isExtendable } from "./isExtendable";
+export { isRegExp } from "./isRegExp";
+export { isDate } from "./isDate";